requirements.txt
numpy==1.24.3
torch==2.0.1
torchvision==0.15.2
opencv-python==4.7.0.72
matplotlib==3.7.1
pandas==2.0.1
scikit-learn==1.2.2
albumentations==1.3.0
pillow==9.5.0
tqdm==4.65.0
tensorboard==2.13.0 